+++ sshd-core/src/test/java/org/apache/sshd/common/session/helpers/AbstractSessionTest.java
@@ -460,7 +460,7 @@ public class AbstractSessionTest extends BaseTestSupport {
         }
 
         @Override
-        public void startService(String name) throws Exception {
+        public void startService(String name, Buffer buffer) throws Exception {
             // ignored
         }