+++ CHANGES.md
@@ -7,7 +7,13 @@
 * The `ChannelSession` provides a mechanism for supporting non-standard extended data (a.k.a. STDERR data)
 in a similar manner as the "regular" data. Please read the relevant section in the main documentation page.
 
+* The user can use a registered `SessionDisconnectHandler` in order be informed and also intervene in cases
+where the code decides to disconnect the session due to various protocol or configuration parameters violations.
+
 ## Behavioral changes and enhancements
 
 * [SSHD-882](https://issues.apache.org/jira/browse/SSHD-882) - Provide hooks to allow users to register a consumer
 for STDERR data sent via the `ChannelSession` - especially for the SFTP subsystem.
+
+* [SSHD=892](https://issues.apache.org/jira/browse/SSHD-882) - Inform user about possible session disconnect prior
+to disconnecting and allow intervention via `SessionDisconnectHandler`.